♻️ init Google認証の初期設定をcommonに移動
